factor out functionality from ExprHelpers
